chore(berm): kill orphan functions
